Migrate TODOs to GitHub Issues
Note that not everything was added to GitHub Issues; this is because
those items were deemed to not be a good fit for Issues and should be
tracked elsewhere. e.g. the list of potential packages to consider using
in LifeLog did not become an Issue because it is not an issue, it's not
an actionable task for development.